+++ src/test_mate/statistic_server/push_data.clj
@@ -32,10 +32,12 @@
     1.0
     (double (/ (:covered pushed-data) (:lines pushed-data)))))
 
-(defn- descreasing-coverage? [push-stats project-name]
-  (let [latest (load-latest-data project-name)]
-    (< (percentage-of-pushed push-stats)
-       (- (:percentage (:overall-coverage latest)) fail-epsilon))))
+(defn- decreasing-coverage? [push-stats project-name]
+  (let [latest (load-latest-data project-name)
+        provided-percentage (percentage-of-pushed push-stats)
+        pushed-percentage (:percentage (:overall-coverage latest))]
+    (println (format "Stats-Server: %s Provided: %s" (* 100 pushed-percentage) (* 100 provided-percentage)))
+    (< pushed-percentage (- provided-percentage fail-epsilon))))
 
 (defn- post-data-to-server [coverage-stats project-name]
   (let [coverage (select-keys coverage-stats [:covered :lines])
@@ -46,13 +48,15 @@
                                       :headers {"auth-token" (config/publish-auth-token)}})
     (println "Successfully pushed " data " to " publish-coverage-url)))
 
+(defn- validate-data [coverage-stats project-name]
+  (when (decreasing-coverage? coverage-stats project-name))
+    (println "Decreasing coverage detected, not allowed by configuration, terminating.")
+    (exit/terminate -1))
+
 (defn- send-data [coverage-stats project-name]
-  (if (and (not (config/allow-decreasing-coverage))
-           (descreasing-coverage? coverage-stats project-name))
-      (do
-         (println "Decreasing coverage detected, not allowed by configuration, terminating.")
-         (exit/terminate -1))
-      (post-data-to-server coverage-stats project-name)))
+  (when-not (config/allow-decreasing-coverage)
+    (validate-data coverage-stats project-name))
+  (post-data-to-server coverage-stats project-name))
 
 (defn- put-project [project-def]
   (let [put-data (only-project-data project-def)]
@@ -73,10 +77,19 @@
   (when (not (= (:status result) :skipped))
     (println (:status result) " -> "(:project-def result))))
 
-(defn publish-statistic-data [coverage-file project-data]
+(defn- read-with-prj-defaults
+  "Reads a project identifier string and constructs a project object with defaults filled in if omitted."
+  [project-data]
   (let [project-def-supplied (only-project-data (edn/read-string project-data))]
-    (send-data (parse/stats coverage-file)
-               (merge (config/default-project) project-def-supplied))))
+    (merge (config/default-project) project-def-supplied)))
+
+(defn publish-statistic-data [coverage-file project-data]
+  (send-data (parse/stats coverage-file) (read-with-prj-defaults project-data)))
+
+(defn check-coverage
+  "checks coverage on server and fails if higher than provided"
+  [coverage-file project-data]
+  (validate-data (parse/stats coverage-file) (read-with-prj-defaults project-data)))
 
 (defn add-project [project-file]
   (let [projects-to-add (edn/read-string (slurp project-file))]
@@ -87,4 +100,5 @@
   (condp = cmd
     "publish-coverage" (publish-statistic-data (first args) (second args))
     "add-project" (add-project (first args))
+    "check-coverage" (check-coverage (first args) (second args))
     (command/exit-with-usage (str "unknown statistic-server command: " cmd) "statistic-server")))