+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- added statistic-server check-coverage for independent checking of coverage diff with server
 
 ## [0.9.0] - 2017-03-23
 ### Added